feat: Restructure files and packages
