+++ src/main/kotlin/io/github/enyason/client/Replicate.kt
@@ -3,12 +3,12 @@ package io.github.enyason.client
 import com.google.gson.reflect.TypeToken
 import io.github.enyason.base.ReplicateConfig
 import io.github.enyason.client.polling.PredictionPollingStrategy
-import io.github.enyason.domain.models.Prediction
-import io.github.enyason.domain.models.isCanceled
-import io.github.enyason.domain.models.isCompleted
-import io.github.enyason.predictable.Predictable
-import io.github.enyason.predictable.validate
-import io.github.enyason.predictable.validateId
+import io.github.enyason.domain.predictions.models.Prediction
+import io.github.enyason.domain.predictions.models.isCanceled
+import io.github.enyason.domain.predictions.models.isCompleted
+import io.github.enyason.predictions.predictable.Predictable
+import io.github.enyason.predictions.predictable.validate
+import io.github.enyason.predictions.predictable.validateId
 import io.github.enyason.predictions.PredictionsApi
 import io.github.enyason.predictions.models.PredictionDTO
