+++ src/main/kotlin/io/github/enyason/domain/predictions/models/PredictionStatus.kt
@@ -1,4 +1,4 @@
-package io.github.enyason.domain.models
+package io.github.enyason.domain.predictions.models
 
 enum class PredictionStatus {
     STARTING,
