+++ src/main/kotlin/io/github/enyason/domain/predictions/models/Urls.kt
@@ -1,4 +1,4 @@
-package io.github.enyason.domain.models
+package io.github.enyason.domain.predictions.models
 
 data class Urls(
     var cancel: String? = null,
