+++ src/main/kotlin/io/github/enyason/predictions/PredictionsApi.kt
@@ -4,8 +4,8 @@ import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import io.github.enyason.base.ReplicateConfig
 import io.github.enyason.base.RetrofitFactory
-import io.github.enyason.domain.mappers.toPrediction
-import io.github.enyason.domain.models.Prediction
+import io.github.enyason.domain.predictions.toPrediction
+import io.github.enyason.domain.predictions.models.Prediction
 import io.github.enyason.predictions.models.PredictionDTO
 import io.github.enyason.predictions.models.toModel
 import okhttp3.ResponseBody
