+++ src/test/kotlin/io/github/enyason/client/ReplicateTest.kt
@@ -1,7 +1,7 @@
 package io.github.enyason.client
 
-import io.github.enyason.domain.models.Prediction
-import io.github.enyason.predictable.Predictable
+import io.github.enyason.domain.predictions.models.Prediction
+import io.github.enyason.predictions.predictable.Predictable
 import io.github.enyason.predictions.PredictionsApi
 import io.mockk.coEvery
 import io.mockk.every
